Fix grid view design.
